+++ h2/connection.py
@@ -11,6 +11,7 @@ from hyperframe.exceptions import InvalidPaddingError
 from hyperframe.frame import (
     GoAwayFrame, WindowUpdateFrame, HeadersFrame, DataFrame, PingFrame,
     PushPromiseFrame, SettingsFrame, RstStreamFrame, PriorityFrame,
+    ContinuationFrame
 )
 from hpack.hpack import Encoder, Decoder
 
@@ -286,6 +287,7 @@ class H2Connection(object):
             RstStreamFrame: self._receive_rst_stream_frame,
             PriorityFrame: self._receive_priority_frame,
             GoAwayFrame: self._receive_goaway_frame,
+            ContinuationFrame: self._receive_naked_continuation,
         }
 
     def _prepare_for_sending(self, frames):
@@ -878,6 +880,7 @@ class H2Connection(object):
         """
         events = []
         self.incoming_buffer.add_data(data)
+        self.incoming_buffer.max_frame_size = self.max_inbound_frame_size
 
         try:
             for frame in self.incoming_buffer:
@@ -903,12 +906,6 @@ class H2Connection(object):
            Removed from the public API.
         """
         try:
-            if frame.body_len > self.max_inbound_frame_size:
-                raise FrameTooLargeError(
-                    "Received overlong frame: length %d, max %d" %
-                    (frame.body_len, self.max_inbound_frame_size)
-                )
-
             # I don't love using __class__ here, maybe reconsider it.
             frames, events = self._frame_dispatch_table[frame.__class__](frame)
         except StreamClosedError as e:
@@ -1140,6 +1137,17 @@ class H2Connection(object):
         events.append(new_event)
 
         return [], events
+
+    def _receive_naked_continuation(self, frame):
+        """
+        A naked CONTINUATION frame has been received. This is always an error,
+        but the type of error it is depends on the state of the stream and must
+        transition the state of the stream, so we need to pass it to the
+        appropriate stream.
+        """
+        stream = self._get_stream_by_id(frame.stream_id)
+        stream.receive_continuation()
+        assert False, "Should not be reachable"
 
     def _local_settings_acked(self):
         """